+++ app/src/main/java/woowacourse/shopping/domain/model/cart/SelectedCartOrder.kt
@@ -1,9 +1,6 @@
 package woowacourse.shopping.domain.model.cart
 
 import kotlinx.serialization.Serializable
-import woowacourse.shopping.domain.model.coupon.Coupon
-import woowacourse.shopping.domain.model.order.DEFAULT_DELIVERY_FEE
-import woowacourse.shopping.domain.model.order.OrderPriceSummary
 
 @Serializable
 data class SelectedCartOrder(
@@ -18,23 +15,6 @@ data class SelectedCartOrder(
             .filter { item -> item.quantity >= quantity }
             .maxOfOrNull { item -> item.price.toLong() }
             ?: 0
-
-    fun calculatePriceSummary(
-        selectedCoupon: Coupon? = null,
-        defaultDeliveryFee: Long = DEFAULT_DELIVERY_FEE,
-    ): OrderPriceSummary {
-        val orderAmount = totalOrderAmount()
-        val couponDiscount = selectedCoupon?.discountAmountFor(this)?.coerceAtMost(orderAmount) ?: 0
-        val deliveryFee = selectedCoupon?.deliveryFeeFor(orderAmount, defaultDeliveryFee) ?: defaultDeliveryFee
-        val totalPaymentPrice = (orderAmount - couponDiscount + deliveryFee).coerceAtLeast(0)
-
-        return OrderPriceSummary(
-            orderAmount = orderAmount,
-            couponDiscount = couponDiscount,
-            deliveryFee = deliveryFee,
-            totalPaymentPrice = totalPaymentPrice,
-        )
-    }
 }
 
 @Serializable
